// dashboard/ui/src/api/generated/models/fileInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */

export interface FileInfo {
  name: string;
  /** @minimum 0 */
  size: number;
  path: string;
  url: string;
}